chatbot/serve_rag/requirements.txt
dspy-ai	
flask
gunicorn
python-dotenv
